allow user and administrator to install parts library
Fritzing may be installed in a location that is read-only but the user
may want to control their own parts library e.g. to keep up-to-date
using the new Git facility.

Look for parts directory in the order:
~/.local/share/fritzing - user override
/usr/local/share/fritzing - adminstrator override
<m_appPath> - the location of translations and help directories, could
be either of the above but as installed will default to PKGDATADIR, may
be set at runtime with -f

Default to <m_appPath>/parts

Use the definition of PKGDATADIR as an indicator that Fritzing will be
installed.